Remove Redis::Distributed.
