--- test/distributed_commands_on_hashes_test.rb
@@ -1,8 +1,0 @@
-require_relative "helper"
-require_relative "lint/hashes"
-
-class TestDistributedCommandsOnHashes < Test::Unit::TestCase
-
-  include Helper::Distributed
-  include Lint::Hashes
-end
